fix indentation to match the default Typespec code style
Typespec has a default code style. Previously, everything was indented by
`typespec-ts-mode-indent-offset`, except for some symbols. Requiring the code
formatter to reformat the code on trivial changes. Now most elements are
indented correctly, including comments. The indentation is implemented such that
the formatter does not necessary has to reformat the code.

The main motivation of this change was to fix the behavior where the following
was indented wrong (where | is the cursor):

  @doc("...")|
  model Test {

Would be indented the following, if inserting a new line (pressing enter):

  @doc("...")
    |
  model Test {

Another issue was:

  @doc("...")
  model Test {|

Was indented into

  @doc("...")
    model Test {
    |

This is fixed by specifying each element type separately.